requirements.txt
discord.py[voice]>=2.0.0
ytmusicapi>=1.8.2
yt_dlp>=2024.10.7